requirements.txt
flask==3.1.0
pydantic==2.10.6
prometheus-client==0.21.1
PyYAML==6.0.2
flask-limiter==3.10.1
flasgger==0.9.7.1
numpy==2.0.0
joblib==1.4.2
werkzeug==3.1.3
pyjwt==2.8.0
scikit-learn==1.5.2
gunicorn
pytest
pandas==2.2.3
pyarrow==18.1.0
evidently==0.5.1
kafka-python==2.0.2
